conda-requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.9.2
aws-msk-iam-sasl-signer-python=1.0.1=pypi_0
boto=2.49.0=py310hca03da5_0
boto3=1.35.49=pypi_0
botocore=1.35.49=pypi_0
brotli-python=1.0.9=py310h313beb8_8
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.9.24=hca03da5_0
click=8.1.7=py310hca03da5_0
confluent-kafka=2.6.1=pypi_0
jmespath=1.0.1=py310hca03da5_0
kafka-python=2.0.2=pypi_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_1
ncurses=6.4=h313beb8_0
openssl=3.0.15=h80987f9_0
pip=24.3.1=pypi_0
pysocks=1.7.1=py310hca03da5_0
python=3.10.15=hb885b13_1
python-dateutil=2.9.0post0=py310hca03da5_2
python-dotenv=0.21.0=py310hca03da5_0
readline=8.2=h1a28f6b_0
s3transfer=0.10.1=py310hca03da5_0
setuptools=75.1.0=py310hca03da5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.45.3=h80987f9_0
tk=8.6.14=h6ba3021_0
tzdata=2024b=h04d1e81_0
urllib3=2.2.3=py310hca03da5_0
wheel=0.44.0=py310hca03da5_0
xz=5.4.6=h80987f9_1
zlib=1.2.13=h18a0788_1
